fix: resolve bug audit findings and switch sync delegate waiting to events

+++ src/tools/lsp/find-references/types.ts
@@ -2,8 +2,8 @@ import { z } from "zod";
 
 export const lspFindReferencesArgsSchema = z.object({
   filePath: z.string(),
-  line: z.number().min(1).describe("1-based"),
-  character: z.number().min(0).describe("0-based"),
+  line: z.number().int().min(1).describe("1-based"),
+  character: z.number().int().min(0).describe("0-based"),
   includeDeclaration: z.boolean().optional().describe("Include the declaration itself"),
 });
